odb: get rid of `the_repository` when handling submodule alternates
The "--recursive" flag for git-grep(1) allows users to grep for a string
across submodule boundaries. To make this work we add each submodule's
object alternate to our own object database so that the objects can be
accessed directly.

The infrastructure for this depends on a global string list of submodule
paths. The caller is expected to call `add_submodule_odb_by_path()` for
each alternate and the object database will then eventually register all
submodule alternates via `do_oid_object_info_extended()` in case it
isn't able to look up a specific object.

This reliance on global state is of course suboptimal with regards to
our libification efforts.

Refactor the logic so that the list of submodule alternates is instead
tracked in the object database itself. This allows us to lose the
condition of `r == the_repository` before registering submodule
alternates as we only ever add submodule alternates to `the_repository`
anyway. As such, behaviour before and after this refactoring should
always be the same.

Rename the functions accordingly.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>